pyproject.toml
[project]
name = "etl-api-crm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "mkdocs>=1.6.1",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.47.0",
    "uvicorn>=0.35.0",
]
